==> 03_data_cleaning.Rmd <==
---
title: "03. Data Cleaning v2"
author: Jessica Guezen
output: pdf_document
---

# Set up

Open CSVs:
```{r}
trait_dat = read.csv("~/Desktop/clim_land_arthropods/Trait_Data.csv",
                     header = TRUE)
species_dat = read.csv("~/Desktop/clim_land_arthropods/Species_Data.csv",
                       header = TRUE)
site_clim_dat = read.csv("~/Desktop/clim_land_arthropods/Site_Clim_Data.csv", header = TRUE)
land_dat = read.csv("~/Desktop/clim_land_arthropods/Landscape_Data.csv",
                    header = TRUE)
study_metdat = read.csv("~/Desktop/clim_land_arthropods/Study_Metadata.csv",
                        header = TRUE)
```

Load packages:
```{r}
library(tidyverse)
```

# Landscape Data

Change from long to wide format:
```{r}
land_dat_wide <-
  land_dat %>%
  pivot_wider(names_from = Radius_meters,
              names_sep = "_",
              values_from = c(Percent_seminatural,
                              Percent_arable,
                              Percent_forest,
                              Percent_urban,
                              Percent_water,
                              Total_Edge_density_kmperha,
                              Shannon_habitat_diversity_index))
rm(land_dat)
```

# Species, Trait, and Study Metadata

Remove all columns from study_metdat except Study.ID and Functional.groups:
```{r}
study_metdat <-
  study_metdat %>%
  dplyr::select(Study.ID, Functional.groups)
```

Get levels for Functional.groups:
```{r}
levels(as.factor(study_metdat$Functional.groups))
```

Add "Pollinator_study", "Pest_study", "Nat_enemy_study", and "Seed_pred_study" variables:
```{r}
study_metdat <-
  study_metdat %>%
  mutate(Pest_study = case_when(
    Functional.groups == "enemies, pests, pollinators" ~ 1,
    Functional.groups == "enemies, pollinators, pests" ~ 1,
    Functional.groups == "enemies, pests" ~ 1,
    Functional.groups == "pests" ~ 1,
    .default = 0))%>%
   mutate(Pollinator_study = case_when(
    Functional.groups == "enemies, pests, pollinators" ~ 1,
    Functional.groups == "enemies, pollinators, pests" ~ 1,
    Functional.groups == "pollinators" ~ 1,
    Functional.groups == "enemies, pollinators" ~ 1,
    .default = 0))%>%
  mutate(Nat_enemy_study = case_when(
    Functional.groups == "enemies"  ~ 1,
    Functional.groups == "enemies, pests, pollinators" ~ 1,
    Functional.groups == "enemies, pollinators, pests" ~ 1,
    Functional.groups == "enemies, pests" ~ 1,
    Functional.groups == "enemies, pollinators" ~ 1,
    .default = 0))%>%
  mutate(Seed_pred_study = case_when(
    Functional.groups == "seed predators"  ~ 1,
    .default = 0))
```

Combine species data and study metadata:
```{r}
species_dat <-
  species_dat %>%
  left_join(study_metdat,
            join_by(StudyID == Study.ID))
rm(study_metdat)
```

Check for duplicates in trait_dat:
```{r}
trait_dat$SpeciesID[duplicated(trait_dat$SpeciesID)]
```

Get levels for Functional_group when SpeciesID = "Syrphidae"
```{r}
trait_dat %>%
  filter(SpeciesID == "Syrphidae") %>%
  reframe(Functional_group)
```

Get levels for Functional_group when SpeciesID = "Diptera"
```{r}
trait_dat%>%
  filter(SpeciesID == "Diptera") %>%
  reframe(Functional_group)
```

Create new SpeciesID's based on Functional_group for Syrphidae and Diptera:
```{r}
trait_dat <-
  trait_dat %>%
  mutate(SpeciesID = 
           case_when(
             SpeciesID == "Syrphidae" & 
              Functional_group == "larval predator, adult pollinator" ~
               "Syrphidae_pred_poll",
             SpeciesID == "Syrphidae" & 
               Functional_group == "pollinator"~
               "Syrphidae_poll",
             SpeciesID == "Syrphidae" & 
               Functional_group == "predator"~
               "Syrphidae_pred",
             .default = SpeciesID
             ))%>%
  mutate(SpeciesID = 
           case_when(
             SpeciesID == "Diptera" & 
              Functional_group == "pollinator" ~
               "Diptera_poll",
             SpeciesID == "Diptera" & 
               Functional_group == "parasitoid of bees"~
               "Diptera_pred",
             .default = SpeciesID
             ))
```

Create new corrected_names based on Pollinator_study and Nat_Enemy_study for Syrphidae and Diptera:
```{r}
species_dat <-
  species_dat %>%
  mutate(corrected_name = 
           case_when(
             corrected_name == "Syrphidae" & 
              Pollinator_study == "1" & 
               Nat_enemy_study == "1" ~
               "Syrphidae_pred_poll",
             corrected_name == "Syrphidae" & 
               Pollinator_study == "1" & 
               Nat_enemy_study == "0" ~
               "Syrphidae_poll",
             corrected_name == "Syrphidae" & 
               Pollinator_study == "0" & 
               Nat_enemy_study == "1" ~
               "Syrphidae_pred",
             .default = corrected_name
             ))%>%
  mutate(corrected_name = 
           case_when(
             corrected_name == "Diptera" & 
               OrganismID == "Diptera sp. (trapnest parasitoid ie larva)" ~
               "Diptera_pred",
             corrected_name == "Diptera" & 
               OrganismID != "Diptera sp. (trapnest parasitoid ie larva)" ~
               "Diptera_poll",
             .default = corrected_name
             ))
```

Combine species and trait data:
```{r}
species_trait_dat <-
  species_dat %>%
  left_join(trait_dat,
            join_by(corrected_name == SpeciesID))
```

Remove Functional.groups variable:
```{r}
species_trait_dat <-
  species_trait_dat %>%
  dplyr::select(-Functional.groups)
```


# All data combined

Remove columns from site_clim_dat: X, Year, Site_description, Crop_species,
Tilling, Local_diversity, and Other_covariate:
```{r}
site_clim_dat <-
  site_clim_dat %>%
  dplyr::select(-X,
         -Year,
         -Site_description,
         -Crop_species,
         -Tilling,
         -Local_diversity,
         -Other_covariate)
```


```{r}
final_dat <-
  species_trait_dat %>%
  left_join(land_dat_wide,
            join_by(
              StudyID == StudyID,
              StudyID_yr == StudyID_yr,
              SiteID == SiteID))%>%
  left_join(site_clim_dat,
            join_by(
              StudyID == StudyID,
              StudyID_yr == StudyID_yr,
              SiteID == SiteID))
```


```{r}
levels(as.factor(final_dat$Functional_group))
```


Create functional group variables:
```{r}
final_dat <-
  final_dat %>%
  mutate(
    fg_pollinator = case_when(
      Functional_group ==
        "larval non-pest herbivore, adult pollinator" |
        Functional_group ==
        "larval parasitoid, adult pollinator" |
        Functional_group ==
        "larval parasitoid, adult pollinator (larva)" |
        Functional_group ==
        "larval pest herbivore, adult pollinator" |
        Functional_group ==
        "larval pest herbivore, adult pollinator (adults)" |
        Functional_group ==
        "larval predator, adult pollinator" |
        Functional_group ==
        "pollinator" ~ 1,
      .default = 0
    ),
    fg_nat_enemy = case_when(
      Functional_group ==
        "aphid-tender, predator" |
        Functional_group ==
        "cleptoparasite"  |
        Functional_group ==
        "larval parasitoid, adult pollinator" |
        Functional_group ==
        "larval parasitoid, adult pollinator (larva)" |
        Functional_group ==
        "larval predator, adult pollinator" |
        Functional_group ==
        "parasitoid" |
        Functional_group ==
        "parasitoid of bees" |
        Functional_group ==
        "predator" ~ 1,
      .default = 0
    ),
    fg_pest = case_when(
      Functional_group ==
        "larval pest herbivore, adult pollinator" |
        Functional_group ==
        "larval pest herbivore, adult pollinator (adults)"  |
        Functional_group ==
        "pest herbivore" ~ 1,
      .default = 0
    )
  )
```

```{r}
levels(as.factor(final_dat$diet_breadth))
levels(as.factor(final_dat$overwintering_habitat))
levels(as.factor(final_dat$dispersal))
```

```{r}
sum(is.na(final_dat$diet_breadth))
sum(is.na(final_dat$overwintering_habitat))
sum(is.na(final_dat$dispersal))
```

# Table S3

```{r}
Table_S3 <- final_dat %>%
  dplyr::select(
    Family,
    diet_breadth,
    dispersal,
    overwintering_habitat,
    fg_pollinator,
    fg_pest,
    fg_nat_enemy
  ) %>%
  distinct() %>%
  rename(
    Diet = diet_breadth,
    Dispersal = dispersal,
    Habitat = overwintering_habitat
  ) %>% mutate(
    Habitat = case_when(
      Habitat == "natural, margin" ~ "specialist",
      Habitat == "natural, margin, crop" ~ "generalist"),
    Functional_group = case_when(
      fg_pollinator == 1 ~ "pollinator",
      fg_pest == 1 ~ "pest",
      fg_nat_enemy == 1 ~ "natural enemy"
    )
  ) %>% 
  na.omit() %>%
  dplyr::select(-c(fg_pollinator, fg_pest, fg_nat_enemy))

write.csv(Table_S3, "~/Desktop/clim_land_arthropods/Table_S3.csv")
```


# Trait syndromes

Create trait (ts) variables for pollinators:
```{r}
final_dat %>%
  filter(
    Pollinator_study == 1) %>%
  filter(
    fg_pollinator == 1) %>%
  filter(!Management == "organic") %>%
  filter(!(StudyID_yr == "Diaz01" &  Block == "11")) %>%
  count(diet_breadth, dispersal, overwintering_habitat) %>% na.omit
```

```{r}
final_dat <-
  final_dat %>%
  mutate(
    #T1: diet generalists, flight dispersal, habitat specialists
    ts_poll_01 = case_when(
      fg_pollinator == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T2: diet generalists, flight/wind dispersal, habitat specialists
    ts_poll_02 = case_when(
      fg_pollinator == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight/wind" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T3: diet specialists, flight dispersal, habitat specialists
    ts_poll_03 = case_when(
      fg_pollinator == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T4: diet specialists, flight dispersal, habitat generalists
    ts_poll_04 = case_when(
      fg_pollinator == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    )
  )
```

```{r}
final_dat %>%
filter(ts_poll_01 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

```{r}
final_dat %>%
filter(ts_poll_03 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

Create trait (ts) variables for pests:
```{r}
final_dat %>%
  filter(Pest_study == 1) %>%
  filter(fg_pest == 1) %>%
  filter(!Management == "organic") %>%
  filter(!(StudyID_yr == "Diaz01" & Block == "11")) %>%
  count(diet_breadth, dispersal, overwintering_habitat) %>% na.omit
```

```{r}
final_dat <-
  final_dat %>%
  mutate(
    #T1: diet generalists, flight dispersal, habitat specialists
    ts_pest_01 = case_when(
      fg_pest == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T2: diet generalists, flight dispersal, habitat generalists
    ts_pest_02 = case_when(
      fg_pest == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T3: diet generalists, flight/wind dispersal, habitat generalists
    ts_pest_03 = case_when(
      fg_pest == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight/wind" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T4: diet specialists, flight dispersal, habitat specialists
    ts_pest_04 = case_when(
      fg_pest == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T5: diet specialists, flight/wind dispersal, habitat specialists
    ts_pest_05 = case_when(
      fg_pest == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight/wind" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T6: diet specialists, flight/wind dispersal, habitat generalists
    ts_pest_06 = case_when(
      fg_pest == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight/wind" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    )
  )
```

```{r}
final_dat %>%
filter(ts_pest_05 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

```{r}
final_dat %>%
filter(ts_pest_06 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

Create trait (ts) variables for natural enemies:
```{r}
final_dat %>%
  filter(Nat_enemy_study == 1) %>%
  filter(fg_nat_enemy == 1) %>%
  filter(!Management == "organic") %>%
  filter(!(StudyID_yr == "Diaz01" & Block == "11")) %>%
  count(diet_breadth, dispersal, overwintering_habitat) %>% na.omit
```

```{r}
final_dat <-
  final_dat %>%
  mutate(
    #T1: diet generalists, flight dispersal, habitat specialists
    ts_nat_enemy_01 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T2: diet generalists, flight dispersal, habitat generalists
    ts_nat_enemy_02 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T3: diet generalists, ground dispersal, habitat specialists
    ts_nat_enemy_03 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "ground" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T4: diet generalists, ground dispersal, habitat generalists
    ts_nat_enemy_04 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "ground" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T5: diet generalists, wind dispersal, habitat specialists
    ts_nat_enemy_05 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "wind" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T6: diet generalists, wind dispersal, habitat generalists
    ts_nat_enemy_06 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "generalist" &
      dispersal == "wind" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T7: diet specialists, flight dispersal, habitat specialists
    ts_nat_enemy_07 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T8: diet specialists, flight dispersal, habitat generalists
    ts_nat_enemy_08 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight" &
        overwintering_habitat == "natural, margin, crop" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T9: diet specialists, flight/wind dispersal, habitat specialists
    ts_nat_enemy_09 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "specialist" &
      dispersal == "flight/wind" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T10: diet specialists, ground dispersal, habitat specialists
    ts_nat_enemy_10 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "specialist" &
      dispersal == "ground" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    ),
    #T11: diet specialists, wind dispersal, habitat specialists
    ts_nat_enemy_11 = case_when(
      fg_nat_enemy == 1 &
      diet_breadth == "specialist" &
      dispersal == "wind" &
        overwintering_habitat == "natural, margin" ~ 1,
      is.na(diet_breadth) | 
        is.na(dispersal) | 
        is.na(overwintering_habitat) ~ NA,
      .default = 0
    )
  )
```

```{r}
final_dat %>%
filter(ts_nat_enemy_02 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

```{r}
final_dat %>%
filter(ts_nat_enemy_03 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

```{r}
final_dat %>%
filter(ts_nat_enemy_04 == 1) %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

```{r}
final_dat %>%
filter(fg_nat_enemy == 1) %>%
filter(diet_breadth == "specialist") %>%
  group_by(Family) %>%
  summarize(sum(Abundance))
```

# Final data cleaning

Remove organic sites and instance where sites within a Block are represented by more than one value for climate variables:
```{r}
final_dat <-
  final_dat %>%
  filter(!Management == "organic") %>%
  dplyr::select(-Management) %>%
  filter(!(StudyID_yr == "Diaz01" &  Block == "11")) #remove single instance where sites within a Block are represented by more than one value for climate variables (n = 4 obs)
```

Calculate abundance and sampling effort: 
```{r}
final_dat <-
  final_dat %>%
  group_by(StudyID,
           Year,
           SiteID,
           Sampling_method,
           Block)%>%
  mutate(total_abund = sum(Abundance),
         #Functional groups:
         fg_pollinator_abund = 
           sum(fg_pollinator*Abundance),
         fg_nat_enemy_abund = 
           sum(fg_nat_enemy*Abundance),
         fg_pest_abund = 
           sum(fg_pest*Abundance),
         #Traits for pollinators:
         ts_poll_01_abund = 
           sum(ts_poll_01*Abundance),
         ts_poll_02_abund = 
           sum(ts_poll_02*Abundance),
         ts_poll_03_abund = 
           sum(ts_poll_03*Abundance),
         ts_poll_04_abund = 
           sum(ts_poll_04*Abundance),
         #Traits for pests:
         ts_pest_01_abund = 
           sum(ts_pest_01*Abundance),
         ts_pest_02_abund = 
           sum(ts_pest_02*Abundance),
         ts_pest_03_abund = 
           sum(ts_pest_03*Abundance),
         ts_pest_04_abund = 
           sum(ts_pest_04*Abundance),
         ts_pest_05_abund = 
           sum(ts_pest_05*Abundance),
         ts_pest_06_abund = 
           sum(ts_pest_06*Abundance),
         #Traits for natural enemies:
         ts_nat_enemy_01_abund = 
           sum(ts_nat_enemy_01*Abundance),
         ts_nat_enemy_02_abund = 
           sum(ts_nat_enemy_02*Abundance),
         ts_nat_enemy_03_abund = 
           sum(ts_nat_enemy_03*Abundance),
         ts_nat_enemy_04_abund = 
           sum(ts_nat_enemy_04*Abundance),
         ts_nat_enemy_05_abund = 
           sum(ts_nat_enemy_05*Abundance),
         ts_nat_enemy_06_abund = 
           sum(ts_nat_enemy_06*Abundance),
         ts_nat_enemy_07_abund = 
           sum(ts_nat_enemy_07*Abundance),
         ts_nat_enemy_08_abund = 
           sum(ts_nat_enemy_08*Abundance),
         ts_nat_enemy_09_abund = 
           sum(ts_nat_enemy_09*Abundance),
         ts_nat_enemy_10_abund = 
           sum(ts_nat_enemy_10*Abundance),
         ts_nat_enemy_11_abund = 
           sum(ts_nat_enemy_11*Abundance)
         ) %>%
  dplyr::select(-corrected_name,
         -OrganismID,
         -Abundance,
         -Authority,
         -Family,
         -Taxonomic_group,
         -Functional_group,
         -diet_breadth,
         -agricultural_specialist,
         -diet_life_history,
         -overwintering_habitat,
         -dispersal,
         -stratum,
         -ts_poll_01,
         -ts_poll_02,
         -ts_poll_03,
         -ts_poll_04,
         -ts_pest_01,
         -ts_pest_02,
         -ts_pest_03,
         -ts_pest_04,
         -ts_pest_05,
         -ts_pest_06,
         -ts_nat_enemy_01,
         -ts_nat_enemy_02,
         -ts_nat_enemy_03,
         -ts_nat_enemy_04,
         -ts_nat_enemy_05,
         -ts_nat_enemy_06,
         -ts_nat_enemy_07,
         -ts_nat_enemy_08,
         -ts_nat_enemy_09,
         -ts_nat_enemy_10,
         -ts_nat_enemy_11,
         -fg_pollinator,
         -fg_nat_enemy,
         -fg_pest) %>%
  distinct(.keep_all = TRUE) %>%
  ungroup()%>%
  mutate(sampling_effort =
           case_when(
             Abundance_duration > 0 & Nb_censuses > 0 ~
               Abundance_duration*Nb_censuses,
             Abundance_duration > 0 & Nb_censuses == 0 ~
               Abundance_duration,
             Abundance_duration == 0 & Nb_censuses > 0 ~
               Nb_censuses,
             Abundance_duration > 0 & is.na(Nb_censuses) ~
               Abundance_duration,
             is.na(Abundance_duration) & Nb_censuses > 0 ~
               Nb_censuses
             ))
```

Calculate abundances scaled by sampling effort:
```{r}
final_dat <-
  final_dat %>%
  group_by(StudyID, Year) %>%
  mutate(sampling_effort_scaled = sampling_effort / max(sampling_effort)) %>%
  ungroup() %>%
  mutate(
    total_abund_scaled = total_abund / sampling_effort_scaled,
    ts_poll_01_abund_scaled =
      ts_poll_01_abund / sampling_effort_scaled,
    ts_poll_02_abund_scaled =
      ts_poll_02_abund / sampling_effort_scaled,
    ts_poll_03_abund_scaled =
      ts_poll_03_abund / sampling_effort_scaled,
    ts_poll_04_abund_scaled =
      ts_poll_04_abund / sampling_effort_scaled,
    ts_pest_01_abund_scaled =
      ts_pest_01_abund / sampling_effort_scaled,
    ts_pest_02_abund_scaled =
      ts_pest_02_abund / sampling_effort_scaled,
    ts_pest_03_abund_scaled =
      ts_pest_03_abund / sampling_effort_scaled,
    ts_pest_04_abund_scaled =
      ts_pest_04_abund / sampling_effort_scaled,
    ts_pest_05_abund_scaled =
      ts_pest_05_abund / sampling_effort_scaled,
    ts_pest_06_abund_scaled =
      ts_pest_06_abund / sampling_effort_scaled,
    ts_nat_enemy_01_abund_scaled =
      ts_nat_enemy_01_abund / sampling_effort_scaled,
    ts_nat_enemy_02_abund_scaled =
      ts_nat_enemy_02_abund / sampling_effort_scaled,
    ts_nat_enemy_03_abund_scaled =
      ts_nat_enemy_03_abund / sampling_effort_scaled,
    ts_nat_enemy_04_abund_scaled =
      ts_nat_enemy_04_abund / sampling_effort_scaled,
    ts_nat_enemy_05_abund_scaled =
      ts_nat_enemy_05_abund / sampling_effort_scaled,
    ts_nat_enemy_06_abund_scaled =
      ts_nat_enemy_06_abund / sampling_effort_scaled,
    ts_nat_enemy_07_abund_scaled =
      ts_nat_enemy_07_abund / sampling_effort_scaled,
    ts_nat_enemy_08_abund_scaled =
      ts_nat_enemy_08_abund / sampling_effort_scaled,
    ts_nat_enemy_09_abund_scaled =
      ts_nat_enemy_09_abund / sampling_effort_scaled,
    ts_nat_enemy_10_abund_scaled =
      ts_nat_enemy_10_abund / sampling_effort_scaled,
    ts_nat_enemy_11_abund_scaled =
      ts_nat_enemy_11_abund / sampling_effort_scaled,
    fg_pollinator_abund_scaled =
      fg_pollinator_abund / sampling_effort_scaled,
    fg_nat_enemy_abund_scaled =
      fg_nat_enemy_abund / sampling_effort_scaled,
    fg_pest_abund_scaled =
      fg_pest_abund / sampling_effort_scaled
  )
```

Rename variables:
```{r}
final_dat <-
  final_dat %>%
  rename(
    c(
      semi_100 = Percent_seminatural_100,
      semi_250 = Percent_seminatural_250,
      semi_500 = Percent_seminatural_500,
      semi_1000 = Percent_seminatural_1000,
      semi_2000 = Percent_seminatural_2000,
      semi_3000 = Percent_seminatural_3000,
      forest_100 = Percent_forest_100,
      forest_250 = Percent_forest_250,
      forest_500 = Percent_forest_500,
      forest_1000 = Percent_forest_1000,
      forest_2000 = Percent_forest_2000,
      forest_3000 = Percent_forest_3000,
      edge_100 = Total_Edge_density_kmperha_100,
      edge_250 = Total_Edge_density_kmperha_250,
      edge_500 = Total_Edge_density_kmperha_500,
      edge_1000 = Total_Edge_density_kmperha_1000,
      edge_2000 = Total_Edge_density_kmperha_2000,
      edge_3000 = Total_Edge_density_kmperha_3000
    )
  )
```

Add variable "semifor" combining % forest and % semi-natural habitat:
```{r}
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    semifor_100 = sum(c(semi_100, forest_100)),
    semifor_250 = sum(c(semi_250, forest_250)),
    semifor_500 = sum(c(semi_500, forest_500)),
    semifor_1000 = sum(c(semi_1000, forest_1000)),
    semifor_2000 = sum(c(semi_2000, forest_2000)),
    semifor_3000 = sum(c(semi_3000, forest_3000))
    )
```

# Climate variables

Calculate climate variables for present conditions:
```{r}
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    present_MAT = case_when(
      Year == "2001" ~
        mean(c_across(ends_with("_temp_01"))),
      Year == "2002" ~
        mean(c_across(ends_with("_temp_02"))),
      Year == "2003" ~
        mean(c_across(ends_with("_temp_03"))),
      Year == "2004" ~
        mean(c_across(ends_with("_temp_04"))),
      Year == "2005" ~
        mean(c_across(ends_with("_temp_05"))),
      Year == "2006" ~
        mean(c_across(ends_with("_temp_06"))),
      Year == "2007" ~
        mean(c_across(ends_with("_temp_07"))),
      Year == "2008" ~
        mean(c_across(ends_with("_temp_08"))),
      Year == "2009" ~
        mean(c_across(ends_with("_temp_09"))),
      Year == "2010" ~
        mean(c_across(ends_with("_temp_10"))),
      Year == "2011" ~
        mean(c_across(ends_with("_temp_11"))),
      Year == "2012" ~
        mean(c_across(ends_with("_temp_12"))),
      Year == "2013" ~
        mean(c_across(ends_with("_temp_13"))),
      Year == "2014" ~
        mean(c_across(ends_with("_temp_14")))
    ),
    present_VAT = case_when(
      Year == "2001" ~
        max(c_across(ends_with("_temp_01"))) - min(c_across(ends_with("_temp_01"))),
      Year == "2002" ~
        max(c_across(ends_with("_temp_02"))) - min(c_across(ends_with("_temp_02"))),
      Year == "2003" ~
        max(c_across(ends_with("_temp_03"))) - min(c_across(ends_with("_temp_03"))),
      Year == "2004" ~
        max(c_across(ends_with("_temp_04"))) - min(c_across(ends_with("_temp_04"))),
      Year == "2005" ~
        max(c_across(ends_with("_temp_05"))) - min(c_across(ends_with("_temp_05"))),
      Year == "2006" ~
        max(c_across(ends_with("_temp_06"))) - min(c_across(ends_with("_temp_06"))),
      Year == "2007" ~
        max(c_across(ends_with("_temp_07"))) - min(c_across(ends_with("_temp_07"))),
      Year == "2008" ~
        max(c_across(ends_with("_temp_08"))) - min(c_across(ends_with("_temp_08"))),
      Year == "2009" ~
        max(c_across(ends_with("_temp_09"))) - min(c_across(ends_with("_temp_09"))),
      Year == "2010" ~
        max(c_across(ends_with("_temp_10"))) - min(c_across(ends_with("_temp_10"))),
      Year == "2011" ~
        max(c_across(ends_with("_temp_11"))) - min(c_across(ends_with("_temp_11"))),
      Year == "2012" ~
        max(c_across(ends_with("_temp_12"))) - min(c_across(ends_with("_temp_12"))),
      Year == "2013" ~
        max(c_across(ends_with("_temp_13"))) - min(c_across(ends_with("_temp_13"))),
      Year == "2014" ~
        max(c_across(ends_with("_temp_14"))) - min(c_across(ends_with("_temp_14")))
    )
  )
```

Calculate climate variables for present 3-year conditions:
```{r}
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    #create MAT variable for each year (1999-2014)
    present3_MAT_99 = mean(c_across(ends_with("_temp_99"))),
    present3_MAT_00 =  mean(c_across(ends_with("_temp_00"))),
    present3_MAT_01 =  mean(c_across(ends_with("_temp_01"))),
    present3_MAT_02 =  mean(c_across(ends_with("_temp_02"))),
    present3_MAT_03 =  mean(c_across(ends_with("_temp_03"))),
    present3_MAT_04 =  mean(c_across(ends_with("_temp_04"))),
    present3_MAT_05 =  mean(c_across(ends_with("_temp_05"))),
    present3_MAT_06 =  mean(c_across(ends_with("_temp_06"))),
    present3_MAT_07 =  mean(c_across(ends_with("_temp_07"))),
    present3_MAT_08 =  mean(c_across(ends_with("_temp_08"))),
    present3_MAT_09 =  mean(c_across(ends_with("_temp_09"))),
    present3_MAT_10 =  mean(c_across(ends_with("_temp_10"))),
    present3_MAT_11 =  mean(c_across(ends_with("_temp_11"))),
    present3_MAT_12 =  mean(c_across(ends_with("_temp_12"))),
    present3_MAT_13 =  mean(c_across(ends_with("_temp_13"))),
    present3_MAT_14 =  mean(c_across(ends_with("_temp_14"))),
    #create VAT variable for each year (1999-2014)
    present3_VAT_99 = max(c_across(ends_with("_temp_99"))) - min(c_across(ends_with("_temp_99"))),
    present3_VAT_00 =  max(c_across(ends_with("_temp_00"))) - min(c_across(ends_with("_temp_00"))),
    present3_VAT_01 =  max(c_across(ends_with("_temp_01"))) - min(c_across(ends_with("_temp_01"))),
    present3_VAT_02 =  max(c_across(ends_with("_temp_02"))) - min(c_across(ends_with("_temp_02"))),
    present3_VAT_03 =  max(c_across(ends_with("_temp_03"))) - min(c_across(ends_with("_temp_03"))),
    present3_VAT_04 =  max(c_across(ends_with("_temp_04"))) - min(c_across(ends_with("_temp_04"))),
    present3_VAT_05 =  max(c_across(ends_with("_temp_05"))) - min(c_across(ends_with("_temp_05"))),
    present3_VAT_06 =  max(c_across(ends_with("_temp_06"))) - min(c_across(ends_with("_temp_06"))),
    present3_VAT_07 =  max(c_across(ends_with("_temp_07"))) - min(c_across(ends_with("_temp_07"))),
    present3_VAT_08 =  max(c_across(ends_with("_temp_08"))) - min(c_across(ends_with("_temp_08"))),
    present3_VAT_09 =  max(c_across(ends_with("_temp_09"))) - min(c_across(ends_with("_temp_09"))),
    present3_VAT_10 =  max(c_across(ends_with("_temp_10"))) - min(c_across(ends_with("_temp_10"))),
    present3_VAT_11 =  max(c_across(ends_with("_temp_11"))) - min(c_across(ends_with("_temp_11"))),
    present3_VAT_12 =  max(c_across(ends_with("_temp_12"))) - min(c_across(ends_with("_temp_12"))),
    present3_VAT_13 =  max(c_across(ends_with("_temp_13"))) - min(c_across(ends_with("_temp_13"))),
    present3_VAT_14 =  max(c_across(ends_with("_temp_14"))) - min(c_across(ends_with("_temp_14")))
)

final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
   #create present3_MAT variable for average over last 3 years
    present3_MAT = case_when(
      Year == "2001" ~ mean(c(
        present3_MAT_99, 
        present3_MAT_00,
        present3_MAT_01)),
      Year == "2002" ~ mean(c( 
        present3_MAT_00,
        present3_MAT_01,
        present3_MAT_02)),
      Year == "2003" ~ mean(c( 
        present3_MAT_01,
        present3_MAT_02,
        present3_MAT_03)),
      Year == "2004" ~ mean(c( 
        present3_MAT_02,
        present3_MAT_03,
        present3_MAT_04)),
      Year == "2005" ~ mean(c( 
        present3_MAT_03,
        present3_MAT_04,
        present3_MAT_05)),
      Year == "2006" ~ mean(c( 
        present3_MAT_04,
        present3_MAT_05,
        present3_MAT_06)),
      Year == "2007" ~ mean(c( 
        present3_MAT_05,
        present3_MAT_06,
        present3_MAT_07)),
      Year == "2008" ~ mean(c( 
        present3_MAT_06,
        present3_MAT_07,
        present3_MAT_08)),
      Year == "2009" ~ mean(c( 
        present3_MAT_07,
        present3_MAT_08,
        present3_MAT_09)),
      Year == "2010" ~ mean(c( 
        present3_MAT_08,
        present3_MAT_09,
        present3_MAT_10)),
      Year == "2011" ~ mean(c( 
        present3_MAT_09,
        present3_MAT_10,
        present3_MAT_11)),
      Year == "2012" ~ mean(c( 
        present3_MAT_10,
        present3_MAT_11,
        present3_MAT_12)),
      Year == "2013" ~ mean(c( 
        present3_MAT_11,
        present3_MAT_12,
        present3_MAT_13)),
      Year == "2014" ~ mean(c( 
        present3_MAT_12,
        present3_MAT_13,
        present3_MAT_14))
      ),
   #create present3_VAT variable for average over last 3 years
    present3_VAT = case_when(
      Year == "2001" ~ mean(c(
        present3_VAT_99, 
        present3_VAT_00,
        present3_VAT_01)),
      Year == "2002" ~ mean(c( 
        present3_VAT_00,
        present3_VAT_01,
        present3_VAT_02)),
      Year == "2003" ~ mean(c( 
        present3_VAT_01,
        present3_VAT_02,
        present3_VAT_03)),
      Year == "2004" ~ mean(c( 
        present3_VAT_02,
        present3_VAT_03,
        present3_VAT_04)),
      Year == "2005" ~ mean(c( 
        present3_VAT_03,
        present3_VAT_04,
        present3_VAT_05)),
      Year == "2006" ~ mean(c( 
        present3_VAT_04,
        present3_VAT_05,
        present3_VAT_06)),
      Year == "2007" ~ mean(c( 
        present3_VAT_05,
        present3_VAT_06,
        present3_VAT_07)),
      Year == "2008" ~ mean(c( 
        present3_VAT_06,
        present3_VAT_07,
        present3_VAT_08)),
      Year == "2009" ~ mean(c( 
        present3_VAT_07,
        present3_VAT_08,
        present3_VAT_09)),
      Year == "2010" ~ mean(c( 
        present3_VAT_08,
        present3_VAT_09,
        present3_VAT_10)),
      Year == "2011" ~ mean(c( 
        present3_VAT_09,
        present3_VAT_10,
        present3_VAT_11)),
      Year == "2012" ~ mean(c( 
        present3_VAT_10,
        present3_VAT_11,
        present3_VAT_12)),
      Year == "2013" ~ mean(c( 
        present3_VAT_11,
        present3_VAT_12,
        present3_VAT_13)),
      Year == "2014" ~ mean(c( 
        present3_VAT_12,
        present3_VAT_13,
        present3_VAT_14))
      )
  )
```

Calculate climate variables for historical conditions:
```{r}
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    #create MAT variable for each baseline year (1961-1980)
    past_MAT_61 = mean(c_across(ends_with("_temp_61"))),
    past_MAT_62 = mean(c_across(ends_with("_temp_62"))),
    past_MAT_63 = mean(c_across(ends_with("_temp_63"))),
    past_MAT_64 = mean(c_across(ends_with("_temp_64"))),
    past_MAT_65 = mean(c_across(ends_with("_temp_65"))),
    past_MAT_66 = mean(c_across(ends_with("_temp_66"))),
    past_MAT_67 = mean(c_across(ends_with("_temp_67"))),
    past_MAT_68 = mean(c_across(ends_with("_temp_68"))),
    past_MAT_69 = mean(c_across(ends_with("_temp_69"))),
    past_MAT_70 = mean(c_across(ends_with("_temp_70"))),
    past_MAT_71 = mean(c_across(ends_with("_temp_71"))),
    past_MAT_72 = mean(c_across(ends_with("_temp_72"))),
    past_MAT_73 = mean(c_across(ends_with("_temp_73"))),
    past_MAT_74 = mean(c_across(ends_with("_temp_74"))),
    past_MAT_75 = mean(c_across(ends_with("_temp_75"))),
    past_MAT_76 = mean(c_across(ends_with("_temp_76"))),
    past_MAT_77 = mean(c_across(ends_with("_temp_77"))),
    past_MAT_78 = mean(c_across(ends_with("_temp_78"))),
    past_MAT_79 = mean(c_across(ends_with("_temp_79"))),
    past_MAT_80 = mean(c_across(ends_with("_temp_80"))),
    #create VAT variable for each baseline year (1961-1980)
    past_VAT_61 = max(c_across(ends_with("_temp_61"))) - min(c_across(ends_with("_temp_61"))),
    past_VAT_62 = max(c_across(ends_with("_temp_62"))) - min(c_across(ends_with("_temp_62"))),
    past_VAT_63 = max(c_across(ends_with("_temp_63"))) - min(c_across(ends_with("_temp_63"))),
    past_VAT_64 = max(c_across(ends_with("_temp_64"))) - min(c_across(ends_with("_temp_64"))),
    past_VAT_65 = max(c_across(ends_with("_temp_65"))) - min(c_across(ends_with("_temp_65"))),
    past_VAT_66 = max(c_across(ends_with("_temp_66"))) - min(c_across(ends_with("_temp_66"))),
    past_VAT_67 = max(c_across(ends_with("_temp_67"))) - min(c_across(ends_with("_temp_67"))),
    past_VAT_68 = max(c_across(ends_with("_temp_68"))) - min(c_across(ends_with("_temp_68"))),
    past_VAT_69 = max(c_across(ends_with("_temp_69"))) - min(c_across(ends_with("_temp_69"))),
    past_VAT_70 = max(c_across(ends_with("_temp_70"))) - min(c_across(ends_with("_temp_70"))),
    past_VAT_71 = max(c_across(ends_with("_temp_71"))) - min(c_across(ends_with("_temp_71"))),
    past_VAT_72 = max(c_across(ends_with("_temp_72"))) - min(c_across(ends_with("_temp_72"))),
    past_VAT_73 = max(c_across(ends_with("_temp_73"))) - min(c_across(ends_with("_temp_73"))),
    past_VAT_74 = max(c_across(ends_with("_temp_74"))) - min(c_across(ends_with("_temp_74"))),
    past_VAT_75 = max(c_across(ends_with("_temp_75"))) - min(c_across(ends_with("_temp_75"))),
    past_VAT_76 = max(c_across(ends_with("_temp_76"))) - min(c_across(ends_with("_temp_76"))),
    past_VAT_77 = max(c_across(ends_with("_temp_77"))) - min(c_across(ends_with("_temp_77"))),
    past_VAT_78 = max(c_across(ends_with("_temp_78"))) - min(c_across(ends_with("_temp_78"))),
    past_VAT_79 = max(c_across(ends_with("_temp_79"))) - min(c_across(ends_with("_temp_79"))),
    past_VAT_80 = max(c_across(ends_with("_temp_80"))) - min(c_across(ends_with("_temp_80")))
  )
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    #create past_MAT variable for average over baseline years
    past_MAT = mean(c_across(starts_with("past_MAT"))),
    #create past_VAT variable for average over baseline years
    past_VAT = mean(c_across(starts_with("past_VAT")))
  )
```

Calculate climate change variables:
```{r}
final_dat <-
  final_dat %>%
  rowwise() %>%
  mutate(
    change_MAT = present_MAT - past_MAT,
    change3_MAT = present3_MAT - past_MAT,
    change_VAT = present_VAT - past_VAT,
    change3_VAT = present3_VAT - past_VAT
  )
```

Add site data:
```{r}
site_dat = read.csv("~/Desktop/PhD/Chapter 1/Data/Working/R/Site_Data.csv",
                    header = TRUE)
```

```{r}
final_dat <-
  final_dat %>%
  left_join(
    site_dat,
    join_by(
      StudyID == StudyID,
      Year == Year,
      StudyID_yr == StudyID_yr,
      Block == Block,
      SiteID == SiteID
    )
  )
```

Remove unnecessary variables:
```{r}
final_dat <- final_dat %>%
  dplyr::select(-ends_with(
    c(
      "_61",
      "_62",
      "_63",
      "_64",
      "_65",
      "_66",
      "_67",
      "_68",
      "_69",
      "_70",
      "_71",
      "_72",
      "_73",
      "_74",
      "_75",
      "_76",
      "_77",
      "_78",
      "_79",
      "_80",
      "_99",
      "_00",
      "_01",
      "_02",
      "_03",
      "_04",
      "_05",
      "_06",
      "_07",
      "_08",
      "_09",
      "_10",
      "_11",
      "_12",
      "_13",
      "_14"
    )
  )) %>%
  dplyr::select(-starts_with(
    c(
      "Jan_",
      "Feb_",
      "Mar_",
      "Apr_",
      "May_",
      "Jun_",
      "Jul_",
      "Aug_",
      "Sep_",
      "Oct_",
      "Nov_",
      "Dec_"
    )
  )) %>%
  dplyr::select(-starts_with(c("Percent_", "Shannon_"))) %>%
  dplyr::select(
    -c(
      "Site_description",
      "Crop_species",
      "Management",
      "Tilling",
      "Local_diversity",
      "Other_covariate"
    )
  )
```

# Write CSV

```{r}
write.csv(final_dat, "~/Desktop/clim_land_arthropods/final_dat.csv")
```
