==> 05_figures.Rmd <==
---
title: "05. Figures"
author: Jessica Guezen
output: pdf_document
---

# Set up

Open csv:
```{r}
climate_land_abund = read.csv(
  "~/Desktop/clim_land_arthropods/final_dat.csv")
```

Load packages:
```{r}
library(arm)
library(lme4)
library(lmerTest)
library(MuMIn)
library(polypoly)
library(psych)
library(performance)
library(car)
library(tidyverse)
library(cowplot)
library(sf)
library(viridis)
library(ggeffects)
library(ggrepel)
library(ggmosaic)
library(ggimage)
library(marginaleffects)
library(glmmTMB)
```

```{r}
climate_land_abund <-
  climate_land_abund %>%
  dplyr::select(-X)
```

# Figure 1

```{r}
tf1a_dat <- data.frame(
  change_MAT = c(-1, 0, 1, -1, 0, 1, -1, 0, 1),
  nat_hab = c(-1, -1, -1, 0, 0, 0, 1, 1, 1))
tf1a_dat <- tf1a_dat %>%
  mutate(
    abundance_1 = 5*nat_hab + 5*change_MAT + 15,
    abundance_2 = 5*nat_hab + 5*change_MAT + 5*(nat_hab*change_MAT) + 15,
    abundance_3 = 5*nat_hab + 5*change_MAT - 5*(nat_hab*change_MAT) + 15,
    abundance_4 = 5*nat_hab - 5*change_MAT + 15,
    abundance_5 = 5*nat_hab - 5*change_MAT + 5*(nat_hab*change_MAT) + 15,
    abundance_6 = 5*nat_hab - 5*change_MAT - 5*(nat_hab*change_MAT) + 15)
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod1a <- lm(
  abundance_1 ~
    nat_hab + change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_1 <-
  c(predict(mod1a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1a <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_1)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_1),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 5/30,
    end = 25/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # axis.title.x = element_blank(),
    axis.line = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Additive",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1a
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod2a <- lm(
  abundance_2 ~
    nat_hab * change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_2 <-
  c(predict(mod2a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1b <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_2)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_2),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 10/30,
    end = 30/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # axis.title.x = element_blank(),
    axis.line = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Synergistic",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1b
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod2a <- lm(
  abundance_3 ~
    nat_hab * change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_3 <-
  c(predict(mod2a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1c <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_3)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_3),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 0,
    end = 20/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # axis.title.x = element_blank(),
    axis.line = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Antagonistic",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1c
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod2a <- lm(
  abundance_4 ~
    nat_hab + change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_4 <-
  c(predict(mod2a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1d <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_4)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_4),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 5/30,
    end = 25/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    axis.line = element_blank(),
    # axis.title.x = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Additive",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1d
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod2a <- lm(
  abundance_5 ~
    nat_hab * change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_5 <-
  c(predict(mod2a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1e <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_5)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_5),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 0,
    end = 20/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    axis.line = element_blank(),
    # axis.title.x = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Synergistic",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1e
```

```{r fig.height=2.5, fig.width=4}
tf1a_dat_hull <- tf1a_dat %>% 
  slice(chull(nat_hab, change_MAT))
tf1a_dat_hull_sf <-
  tf1a_dat_hull %>% st_as_sf(coords = c("nat_hab", "change_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

tf1a_dat_xy <-
  with(tf1a_dat,
       expand.grid(
         nat_hab = seq(min(nat_hab),
                   max(nat_hab),
                   len = 400),
         change_MAT = seq(min(change_MAT),
                        max(change_MAT),
                        len = 400)
       ))

tf1a_dat_xy_sf = st_as_sf(tf1a_dat_xy,
                         coords = c("nat_hab", "change_MAT"))
plot_tf1a_crop <-
  st_intersection(
    tf1a_dat_xy_sf,
    tf1a_dat_hull_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(nat_hab = X,
         change_MAT = Y)

mod2a <- lm(
  abundance_6 ~
    nat_hab * change_MAT,
  data = tf1a_dat
)

plot_tf1a_crop$abundance_6 <-
  c(predict(mod2a, newdata = plot_tf1a_crop, se = FALSE))

heatmap_tf1f <- ggplot() +
  geom_raster(data = plot_tf1a_crop,
              aes(x = nat_hab,
                  y = change_MAT,
                  fill = abundance_6)) +
  geom_contour(
    data = plot_tf1a_crop,
    aes(x = nat_hab,
        y = change_MAT,
        z = abundance_6),
    bins = 10,
    color = 'white',
    alpha = 0.5
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    begin = 10/30,
    end = 30/30,
    alpha = 1,
    breaks = c(5, 15, 25),
    labels = c("low", "med", "high")
    ) +
  scale_x_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  scale_y_continuous(
    breaks = c(-1, 1),
    labels = c("low", "high")) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    axis.line = element_blank(),
    # axis.title.x = element_blank(),
    legend.position = "right",
    # legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")) +
  labs(
    title = "Antagonistic",
    fill = "Abundance",
       x = "nat.hab (%)",
       y = "change in MAT")
heatmap_tf1f
```

# Figure 2

```{r}
library(ggplot2)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
library(ggspatial)
world <- ne_countries(scale = "medium", returnclass = "sf")
```

```{r}
fig2a <- ggplot(data = world) +
  geom_sf() +
  stat_summary_hex(
    data = climate_land_abund,
    aes(x = Longitude,
        y = Latitude,
        z = change3_MAT),
    fun = "mean") +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_bw() +
  coord_sf(
    xlim = c(
      min(climate_land_abund$Longitude) - 2,
      max(climate_land_abund$Longitude) + 2
    ),
    ylim = c(
      min(climate_land_abund$Latitude) - 2,
      max(climate_land_abund$Latitude) + 2
    ),
    expand = FALSE
  ) +
  annotation_scale(location = "br", width_hint = 0.2) +
  annotation_north_arrow(
    location = "tr",
    height = unit(0.65, "cm"),
    width = unit(0.65, "cm"),
    which_north = "true",
    pad_x = unit(0.1, "in"),
    pad_y = unit(0.1, "in"),
    style = north_arrow_minimal
  ) +
  labs(fill = "MAT (°C)")
fig2a
```

```{r}
fig2b <- ggplot(data = world) +
  geom_sf() +
  stat_summary_hex(
    data = climate_land_abund,
    aes(x = Longitude,
        y = Latitude,
        z = change3_VAT),
    fun = "mean") +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_bw() +
  coord_sf(
    xlim = c(
      min(climate_land_abund$Longitude) - 2,
      max(climate_land_abund$Longitude) + 2
    ),
    ylim = c(
      min(climate_land_abund$Latitude) - 2,
      max(climate_land_abund$Latitude) + 2
    ),
    expand = FALSE
  ) +
  annotation_scale(location = "br", width_hint = 0.2) +
  annotation_north_arrow(
    location = "tr",
    height = unit(0.65, "cm"),
    width = unit(0.65, "cm"),
    which_north = "true",
    pad_x = unit(0.1, "in"),
    pad_y = unit(0.1, "in"),
    style = north_arrow_minimal
  ) +
  labs(fill = "VAT (°C)")
fig2b
```

# Figure 3

Pollinators:
```{r}
climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  count(ts_poll_01_abund > 0)
climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  count(ts_poll_02_abund > 0)
climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  count(ts_poll_03_abund > 0)
climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  count(ts_poll_04_abund > 0)
```

Pests:
```{r}
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_01_abund > 0)
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_02_abund > 0)
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_03_abund > 0)
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_04_abund > 0)
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_05_abund > 0)
climate_land_abund %>%
  filter(Pest_study == 1) %>%
  count(ts_pest_06_abund > 0)
```

Natural enemies:
```{r}
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_01_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_02_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_03_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_04_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_05_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_06_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_07_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_08_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_09_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_10_abund > 0)
climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  count(ts_nat_enemy_11_abund > 0)
```




```{r}
Trait_bar <- data.frame(
  abund = c(
    #poll
    0, 0, 0, 0, 
    775, 0, 0, 0,
    193, 0, 0, 0,
    323, 0, 0, 0,
    #pest
    1, 11, 0, 0, 
    15, 0, 0, 0,
    0, 209, 0, 0,
    9, 235, 0, 0,
    #nat_enemy
    496, 0, 519, 119, 
    295, 0, 349, 292,
    198, 0, 0, 0,
    2, 134, 14, 81
  ),
  fg = c(
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pollinators",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "pests",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies",
    "natural enemies"
  ),
  diet = c(
    "generalist", 
    "generalist", 
    "generalist", 
    "generalist",
    "generalist", 
    "generalist", 
    "generalist", 
    "generalist",  
    "specialist", 
    "specialist", 
    "specialist", 
    "specialist", 
    "specialist", 
    "specialist", 
    "specialist", 
    "specialist"),
  habitat = c(
    "habitat generalist",
    "habitat generalist",
    "habitat generalist",
    "habitat generalist",
    "habitat specialist",
    "habitat specialist",
    "habitat specialist",
    "habitat specialist",
    "habitat generalist",
    "habitat generalist",
    "habitat generalist",
    "habitat generalist",
    "habitat specialist",
    "habitat specialist",
    "habitat specialist",
    "habitat specialist"),
  dispersal = c("flight", "flight/wind", "ground", "wind")
)
Trait_bar <- Trait_bar %>%
  mutate(percent = case_when(
    fg == "pollinators" ~ (abund / 1086) * 100,
    fg == "pests" ~ (abund / 902) * 100,
    fg == "natural enemies" ~ (abund / 1728) * 100
  )) 
Trait_bar[Trait_bar == 0] <- NA
```

fig3_plot:
```{r}
fig3_plot <- ggplot(
  data = Trait_bar,
  aes(y = percent,
      x = factor(dispersal, levels = c("wind", "flight/wind", "flight", "ground")),
      fill = diet)) +
  geom_bar(stat = "identity",
           width = 0.8,
           position = position_dodge(width = .8),
           colour = "black") +
  geom_hline(yintercept = 20, 
             linetype = "dashed") +
  facet_grid(
    factor(fg, levels = c("pollinators", "pests", "natural enemies")) ~ habitat, scale = "free_y") +
  labs(y = "% sites present") +
  scale_fill_manual(
    values = c("#f0f921", "#ed7953"), 
    labels = c("diet\ngeneralist", "diet\nspecialist"),
    drop = T) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    legend.title = element_blank(),
    legend.position = "bottom",
    legend.direction = "horizontal",
    axis.title.x = element_blank(),
    axis.text.x = element_text(
      angle = 90,
      vjust = 0.5,
      hjust = 1),
    axis.line = element_blank(),
    panel.grid = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1.1,
      fill = NA),
    strip.background = element_blank(),
    strip.text = element_text(size = 9,
                               colour = "black"),
    plot.margin = unit(c(0,0.1,0,0), "cm")
  )
fig3_plot
```

# Figure 4

```{r}
citable_01 <- read.csv("~/Desktop/clim_land_arthropods/citable_01.csv")
citable_05 <- read.csv("~/Desktop/clim_land_arthropods/citable_05.csv")
citable_06 <- read.csv("~/Desktop/clim_land_arthropods/citable_06.csv")
citable_07 <- read.csv("~/Desktop/clim_land_arthropods/citable_07.csv")
citable_08 <- read.csv("~/Desktop/clim_land_arthropods/citable_08.csv")
citable_09 <- read.csv("~/Desktop/clim_land_arthropods/citable_09.csv")
citable_10 <- read.csv("~/Desktop/clim_land_arthropods/citable_10.csv")
citable_11 <- read.csv("~/Desktop/clim_land_arthropods/citable_11.csv")
```

## Formatting

```{r}
citable_01 <- citable_01 %>%
  mutate(fg_ts = "fg_01")
citable_05 <- citable_05 %>%
  mutate(fg_ts = "fg_05")
citable_06 <- citable_06 %>%
  mutate(fg_ts = "fg_06")
citable_07 <- citable_07 %>%
  mutate(fg_ts = "fg_07")
citable_08 <- citable_08 %>%
  mutate(fg_ts = "fg_08")
citable_09 <- citable_09 %>%
  mutate(fg_ts = "fg_09")
citable_10 <- citable_10 %>%
  mutate(fg_ts = "fg_10")
citable_11 <- citable_11 %>%
  mutate(fg_ts = "fg_11")

ci_graphic <- full_join(
  citable_01,
  citable_05
)
ci_graphic <- full_join(
  ci_graphic,
  citable_06
)
ci_graphic <- full_join(
  ci_graphic,
  citable_07
)
ci_graphic <- full_join(
  ci_graphic,
  citable_08
)
ci_graphic <- full_join(
  ci_graphic,
  citable_09
)
ci_graphic <- full_join(
  ci_graphic,
  citable_10
)
ci_graphic <- full_join(
  ci_graphic,
  citable_11
)
ci_graphic <- ci_graphic %>%
  pivot_longer(
    cols = ends_with("_2.5"),
    names_to = c("spatial_scale"),
    names_pattern = "CI_(.*)_2.5",
    values_to = "CI_2.5"
  ) %>%
  mutate(
    CI_97.5 = case_when(
      spatial_scale == 100 ~ CI_100_97.5,
      spatial_scale == 250 ~ CI_250_97.5,
      spatial_scale == 500 ~ CI_500_97.5,
      spatial_scale == 1000 ~ CI_1000_97.5,
      spatial_scale == 2000 ~ CI_2000_97.5,
      spatial_scale == 3000 ~ CI_3000_97.5
    ),
    Estimate = case_when(
      spatial_scale == 100 ~ Estimate_100,
      spatial_scale == 250 ~ Estimate_250,
      spatial_scale == 500 ~ Estimate_500,
      spatial_scale == 1000 ~ Estimate_1000,
      spatial_scale == 2000 ~ Estimate_2000,
      spatial_scale == 3000 ~ Estimate_3000
    )
  ) %>%
  dplyr::select(-CI_100_97.5,
                -CI_250_97.5,
                -CI_500_97.5,
                -CI_1000_97.5,
                -CI_2000_97.5,
                -CI_3000_97.5,
                -Estimate_100,
                -Estimate_250,
                -Estimate_500,
                -Estimate_1000,
                -Estimate_2000,
                -Estimate_3000
                ) %>%
  mutate(
    significant = case_when(
      CI_2.5 > 0 & CI_97.5 > 0 |
        CI_2.5 < 0 & CI_97.5 < 0 ~ "yes",
      CI_2.5 <= 0 & CI_97.5 >= 0 |
        CI_2.5 >= 0 & CI_97.5 <= 0 ~ "no",
      .default = NA
    )
  )
ci_graphic$spatial_scale <-
  factor(
    ci_graphic$spatial_scale,
    ordered = T,
    levels = c("3000", "2000", "1000", "500", "250", "100")
  )
ci_graphic <- ci_graphic %>%
  mutate(
    term = str_replace_all(term, "cond.scale", ""),
    term = str_replace_all(term, "scale", "")) %>%
  mutate(
    term = case_when(
      term == "(semifor_l)" ~ "nat.hab",
      term == "(change3_MAT)" ~ "MAT",
      term == "(change3_VAT)" ~ "VAT",
      term == "(semifor_l):(change3_MAT)" ~ " nat.hab*MAT",
      term == "(change3_MAT):(semifor_l)" ~ " nat.hab*MAT",
      term == "(semifor_l):(change3_VAT)" ~ " nat.hab*VAT",
      term == "(change3_VAT):(semifor_l)" ~ " nat.hab*VAT",
      term == "(change3_MAT):(change3_VAT)" ~ " MAT*VAT",
      term == "(change3_VAT):(change3_MAT)" ~ " MAT*VAT"
      ))

level_order <- c(
  "nat.hab",
  "MAT",
  "VAT",
  " nat.hab*MAT",
  " nat.hab*VAT",
  " MAT*VAT"
  )
```

## 01. All arthropods

```{r}
ci_01_mods <- ci_graphic %>%
  filter(fg_ts == "fg_01")
```

plot_01_mods:
```{r}
plot_01_mods <-
  ggplot(data = ci_01_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  # scale_x_continuous(breaks = c(-0.8, -0.4, 0, 0.4, 0.8)) +
  labs(x = "95% CI",
       y = "All arthropods",
       title = "0.1-3 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_01_mods
```

## 05. Diet generalist pollinators

```{r}
ci_05_mods <- ci_graphic %>%
  filter(fg_ts == "fg_05" &
           !spatial_scale == "3000")
```

plot_05_mods:
```{r}
plot_05_mods <-
  ggplot(data = ci_05_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "DG pollinators",
       title = "0.1-2 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_05_mods
```

## 06. Diet specialist pollinators

```{r}
ci_06_mods <- ci_graphic %>%
  filter(fg_ts == "fg_06" &
           !spatial_scale == "500" &
           !spatial_scale == "1000" &
           !spatial_scale == "2000" &
           !spatial_scale == "3000")
```

plot_06_mods:
```{r}
plot_06_mods <-
  ggplot(data = ci_06_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "DS pollinators",
       title = "0.1-0.25 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_06_mods
```

## 07. Habitat generalist pests

```{r}
ci_07_mods <- ci_graphic %>%
  filter(fg_ts == "fg_07")
```

plot_07_mods:
```{r}
plot_07_mods <-
  ggplot(data = ci_07_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "HG pests",
       title = "0.1-3 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_07_mods
```

## 08. Habitat specialist pests

```{r}
ci_08_mods <- ci_graphic %>%
  filter(fg_ts == "fg_08" &
           !spatial_scale == "2000" &
           !spatial_scale == "3000")
```

plot_08_mods:
```{r}
plot_08_mods <-
  ggplot(data = ci_08_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "HS pests",
       title = "0.1-1 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_08_mods
```

## 09. Habitat generalist flight dispersal natural enemies

```{r}
ci_09_mods <- ci_graphic %>%
  filter(fg_ts == "fg_09" &
           !spatial_scale == "2000" &
           !spatial_scale == "3000"
         )
```

plot_09_mods:
```{r}
plot_09_mods <-
  ggplot(data = ci_09_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "HG flight nat.en.",
       title = "0.1-1 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_09_mods
```

## 10. Habitat specialist ground dispersal natural enemies

```{r}
ci_10_mods <- ci_graphic %>%
  filter(fg_ts == "fg_10" &
           !spatial_scale == "3000")
```

plot_10_mods:
```{r}
plot_10_mods <-
  ggplot(data = ci_10_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "HS ground nat.en.",
       title = "0.1-2 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_10_mods
```

## 11. Habitat generalist ground dispersal natural enemies

```{r}
ci_11_mods <- ci_graphic %>%
  filter(fg_ts == "fg_11" &
           !spatial_scale == "100" &
           !spatial_scale == "1000" &
           !spatial_scale == "2000" &
           !spatial_scale == "3000")
```

plot_11_mods:
```{r}
plot_11_mods <-
  ggplot(data = ci_11_mods, aes(
    y = factor(term, 
               level = rev(level_order)
               ),
    alpha = spatial_scale,
    colour = factor(significant)
  )) +
  geom_vline(
    aes(xintercept = 0),
    colour = "black",
    linetype = "dashed",
    linewidth = 0.2
  ) +
  geom_errorbarh(aes(xmin = CI_2.5, 
                     xmax = CI_97.5),
                 position = position_dodge(width = 0.6),
                 linewidth = 0.3,
                 height = 0) +
  theme_classic() +
  theme(
    # aspect.ratio = 1,
    axis.title.y = element_blank(),
    axis.line = element_blank(),
    axis.text.y = element_text(
     face = c('plain'),
      colour = c("black")),
    legend.position = "none",
    plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title = element_text(size = 8),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  scale_alpha_manual(values = c(1, 1, 1, 1, 1, 1)) +
  scale_color_manual(values = c("gray80", "#0d0887")) +
  labs(x = "95% CI",
       y = "HG ground nat.en.",
       title = "0.25-0.5 km:") +
  guides(alpha = guide_legend(reverse = TRUE,
                              title = "Spatial scale (m)"))
plot_11_mods
```

# Figure 5

## 01. All arthropods

500 m:
```{r}
climate_land_abund_01_500 <-
  climate_land_abund %>%
  dplyr::select(
    total_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_01 = total_abund,
    semifor_l = semifor_500
  )
```

### heatmap_01.1

```{r}
plot_01_500_dat_hull <- climate_land_abund_01_500 %>%
  slice(chull(semifor_l, change3_MAT))

plot_01_500_dat_hull <- plot_01_500_dat_hull %>% dplyr::select(semifor_l, change3_MAT)

plot_01_500_dat_hull_sf <- plot_01_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_01_500_dat_xy <-
  with(climate_land_abund_01_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_01_500_dat_xy_sf = st_as_sf(
  plot_01_500_dat_xy,
  coords = c("semifor_l", "change3_MAT"))

plot_01_crop <-
  st_intersection(
    plot_01_500_dat_hull_sf,
    plot_01_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_MAT = Y) %>%
  mutate(change3_VAT = mean(climate_land_abund_01_500$change3_VAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_01 <- glmmTMB(
  abund_01 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT)  + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_01_500,
  na.action = "na.fail"
)

plot_01_crop$abund_01 <- c(predict(
  mod_01,
  newdata = plot_01_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_01.1 <- ggplot() +
  geom_raster(data = plot_01_crop,
              aes(x = semifor_l,
                  y = change3_MAT,
                  fill = abund_01)) +
  geom_contour(
    data = plot_01_crop,
    aes(x = semifor_l,
        y = change3_MAT,
        z = abund_01),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "MAT (°C)",
    )
heatmap_01.1
```

### heatmap_01.2

```{r}
plot_01_500_dat_hull <- climate_land_abund_01_500 %>%
  slice(chull(semifor_l, change3_VAT))

plot_01_500_dat_hull <- plot_01_500_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_01_500_dat_hull_sf <- plot_01_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_01_500_dat_xy <-
  with(climate_land_abund_01_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_01_500_dat_xy_sf = st_as_sf(
  plot_01_500_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_01_crop <-
  st_intersection(
    plot_01_500_dat_hull_sf,
    plot_01_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_01_500$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_01 <- glmmTMB(
  abund_01 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT)  + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_01_500,
  na.action = "na.fail"
)

plot_01_crop$abund_01 <- c(predict(
  mod_01,
  newdata = plot_01_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_01.2 <- ggplot() +
  geom_raster(data = plot_01_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_01)) +
  geom_contour(
    data = plot_01_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_01),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "VAT (°C)",
    )
heatmap_01.2
```

## 05. Diet generalist pollinators

500 m:
```{r}
climate_land_abund_05_500 <-
  climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  dplyr::select(
    ts_poll_01_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_05 = ts_poll_01_abund,
    semifor_l = semifor_500
  )
```

### heatmap_05.1

```{r}
plot_05_500_dat_hull <- climate_land_abund_05_500 %>%
  slice(chull(semifor_l, change3_MAT))

plot_05_500_dat_hull <- plot_05_500_dat_hull %>% dplyr::select(semifor_l, change3_MAT)

plot_05_500_dat_hull_sf <- plot_05_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_05_500_dat_xy <-
  with(climate_land_abund_05_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_05_500_dat_xy_sf = st_as_sf(
  plot_05_500_dat_xy,
  coords = c("semifor_l", "change3_MAT"))

plot_05_crop <-
  st_intersection(
    plot_05_500_dat_hull_sf,
    plot_05_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_MAT = Y) %>%
  mutate(change3_VAT = mean(climate_land_abund_05_500$change3_VAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_05 <- glmmTMB(
  abund_05 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_05_500,
  na.action = "na.fail"
)

plot_05_crop$abund_05 <- c(predict(
  mod_05,
  newdata = plot_05_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_05.1 <- ggplot() +
  geom_raster(data = plot_05_crop,
              aes(x = semifor_l,
                  y = change3_MAT,
                  fill = abund_05)) +
  geom_contour(
    data = plot_05_crop,
    aes(x = semifor_l,
        y = change3_MAT,
        z = abund_05),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    breaks = c(10, 20, 30, 40)
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    axis.line = element_blank(),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "MAT (°C)",
    )
heatmap_05.1
```


### heatmap_05.2

```{r}
plot_05_500_dat_hull <- climate_land_abund_05_500 %>%
  slice(chull(semifor_l, change3_VAT))

plot_05_500_dat_hull <- plot_05_500_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_05_500_dat_hull_sf <- plot_05_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_05_500_dat_xy <-
  with(climate_land_abund_05_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_05_500_dat_xy_sf = st_as_sf(
  plot_05_500_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_05_crop <-
  st_intersection(
    plot_05_500_dat_hull_sf,
    plot_05_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_05_500$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_05 <- glmmTMB(
  abund_05 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_05_500,
  na.action = "na.fail"
)

plot_05_crop$abund_05 <- c(predict(
  mod_05,
  newdata = plot_05_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_05.2 <- ggplot() +
  geom_raster(data = plot_05_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_05)) +
  geom_contour(
    data = plot_05_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_05),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    breaks = c(20, 30, 40)
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    axis.line = element_blank(),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "VAT (°C)",
    )
heatmap_05.2
```

## 06. Diet specialist pollinators

100 m:
```{r}
climate_land_abund_06_100 <-
  climate_land_abund %>%
  filter(Pollinator_study == 1) %>%
  dplyr::select(
    ts_poll_03_abund,
    sampling_effort_scaled,
    semifor_100,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_06 = ts_poll_03_abund,
    semifor_l = semifor_100
  )
```

### heatmap_06.2

```{r}
plot_06_100_dat_hull <- climate_land_abund_06_100 %>%
  slice(chull(semifor_l, change3_VAT))

plot_06_100_dat_hull <- plot_06_100_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_06_100_dat_hull_sf <- plot_06_100_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_06_100_dat_xy <-
  with(climate_land_abund_06_100,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_06_100_dat_xy_sf = st_as_sf(
  plot_06_100_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_06_crop <-
  st_intersection(
    plot_06_100_dat_hull_sf,
    plot_06_100_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_06_100$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_06 <- glmmTMB(
  abund_06 ~
    scale(semifor_l) * scale(change3_VAT) + scale(change3_MAT) + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_06_100,
  na.action = "na.fail"
)

plot_06_crop$abund_06 <- c(predict(
  mod_06,
  newdata = plot_06_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_06.2 <- ggplot() +
  geom_raster(data = plot_06_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_06)) +
  geom_contour(
    data = plot_06_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_06),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    breaks = c(7, 14, 21, 28)
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, face = "bold"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.1 km",
       y = "VAT (°C)",
    )
heatmap_06.2
```

## 07. Habitat generalist pests

500 m:
```{r}
climate_land_abund_07_500 <-
  climate_land_abund %>%
  filter(Pest_study == 1) %>%
  dplyr::select(
    ts_pest_06_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_07 = ts_pest_06_abund,
    semifor_l = semifor_500
  )
```

### heatmap_07.1

```{r}
plot_07_500_dat_hull <- climate_land_abund_07_500 %>%
  slice(chull(semifor_l, change3_MAT))

plot_07_500_dat_hull <- plot_07_500_dat_hull %>% dplyr::select(semifor_l, change3_MAT)

plot_07_500_dat_hull_sf <- plot_07_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_07_500_dat_xy <-
  with(climate_land_abund_07_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_07_500_dat_xy_sf = st_as_sf(
  plot_07_500_dat_xy,
  coords = c("semifor_l", "change3_MAT"))

plot_07_crop <-
  st_intersection(
    plot_07_500_dat_hull_sf,
    plot_07_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_MAT = Y) %>%
  mutate(change3_VAT = mean(climate_land_abund_07_500$change3_VAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_07 <- glmmTMB(
  abund_07 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_07_500,
  na.action = "na.fail"
)

plot_07_crop$abund_07 <- c(predict(
  mod_07,
  newdata = plot_07_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_07.1 <- ggplot() +
  geom_raster(data = plot_07_crop,
              aes(x = semifor_l,
                  y = change3_MAT,
                  fill = abund_07)) +
  geom_contour(
    data = plot_07_crop,
    aes(x = semifor_l,
        y = change3_MAT,
        z = abund_07),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "MAT (°C)",
    )
heatmap_07.1
```

### heatmap_07.2

```{r}
plot_07_500_dat_hull <- climate_land_abund_07_500 %>%
  slice(chull(semifor_l, change3_VAT))

plot_07_500_dat_hull <- plot_07_500_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_07_500_dat_hull_sf <- plot_07_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_07_500_dat_xy <-
  with(climate_land_abund_07_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_07_500_dat_xy_sf = st_as_sf(
  plot_07_500_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_07_crop <-
  st_intersection(
    plot_07_500_dat_hull_sf,
    plot_07_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_07_500$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_07 <- glmmTMB(
  abund_07 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_07_500,
  na.action = "na.fail"
)

plot_07_crop$abund_07 <- c(predict(
  mod_07,
  newdata = plot_07_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_07.2 <- ggplot() +
  geom_raster(data = plot_07_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_07)) +
  geom_contour(
    data = plot_07_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_07),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "VAT (°C)",
    )
heatmap_07.2
```

## 08. Habitat specialist pests

500 m:
```{r}
climate_land_abund_08_500 <-
  climate_land_abund %>%
  filter(Pest_study == 1) %>%
  dplyr::select(
    ts_pest_05_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_08 = ts_pest_05_abund,
    semifor_l = semifor_500
  )
```

### heatmap_08.1

```{r}
plot_08_500_dat_hull <- climate_land_abund_08_500 %>%
  slice(chull(semifor_l, change3_MAT))

plot_08_500_dat_hull <- plot_08_500_dat_hull %>% dplyr::select(semifor_l, change3_MAT)

plot_08_500_dat_hull_sf <- plot_08_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_08_500_dat_xy <-
  with(climate_land_abund_08_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_08_500_dat_xy_sf = st_as_sf(
  plot_08_500_dat_xy,
  coords = c("semifor_l", "change3_MAT"))

plot_08_crop <-
  st_intersection(
    plot_08_500_dat_hull_sf,
    plot_08_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_MAT = Y) %>%
  mutate(change3_VAT = mean(climate_land_abund_08_500$change3_VAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_08 <- glmmTMB(
  abund_08 ~
    scale(semifor_l) + scale(change3_VAT) + scale(change3_MAT) + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_08_500,
  na.action = "na.fail"
)

plot_08_crop$abund_08 <- c(predict(
  mod_08,
  newdata = plot_08_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_08.1 <- ggplot() +
  geom_raster(data = plot_08_crop,
              aes(x = semifor_l,
                  y = change3_MAT,
                  fill = abund_08)) +
  geom_contour(
    data = plot_08_crop,
    aes(x = semifor_l,
        y = change3_MAT,
        z = abund_08),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.title.x = element_text(size = 8, colour = "grey50"),
    axis.line = element_blank(),
    axis.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "MAT (°C)",
    )
heatmap_08.1
```

### heatmap_08.3

```{r}
plot_08_500_dat_hull <- climate_land_abund_08_500 %>%
  slice(chull(change3_VAT, change3_MAT))

plot_08_500_dat_hull <- plot_08_500_dat_hull %>% dplyr::select(change3_VAT, change3_MAT)

plot_08_500_dat_hull_sf <- plot_08_500_dat_hull %>%
  st_as_sf(coords = 
             c("change3_VAT",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_08_500_dat_xy <-
  with(climate_land_abund_08_500,
       expand.grid(
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_08_500_dat_xy_sf = st_as_sf(
  plot_08_500_dat_xy,
  coords = c("change3_VAT", "change3_MAT"))

plot_08_crop <-
  st_intersection(
    plot_08_500_dat_hull_sf,
    plot_08_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(change3_VAT = X, 
         change3_MAT = Y) %>%
  mutate(semifor_l = mean(climate_land_abund_08_500$semifor_l),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_08 <- glmmTMB(
  abund_08 ~
    scale(semifor_l) + scale(change3_VAT) + scale(change3_MAT) + (1 | StudyID_yr) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_08_500,
  na.action = "na.fail"
)

plot_08_crop$abund_08 <- c(predict(
  mod_08,
  newdata = plot_08_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_08.3 <- ggplot() +
  geom_raster(data = plot_08_crop,
              aes(x = change3_VAT,
                  y = change3_MAT,
                  fill = abund_08)) +
  geom_contour(
    data = plot_08_crop,
    aes(x = change3_VAT,
        y = change3_MAT,
        z = abund_08),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.title.x = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "VAT (°C)",
       y = "MAT (°C)",
    )
heatmap_08.3
```



## 09. Habitat generalist, flight natural enemies

500 m:
```{r}
climate_land_abund_09_500 <-
  climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  dplyr::select(
    ts_nat_enemy_02_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_09 = ts_nat_enemy_02_abund,
    semifor_l = semifor_500
  )
```

### heatmap_09.2

```{r}
plot_09_500_dat_hull <- climate_land_abund_09_500 %>%
  slice(chull(semifor_l, change3_VAT))

plot_09_500_dat_hull <- plot_09_500_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_09_500_dat_hull_sf <- plot_09_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_09_500_dat_xy <-
  with(climate_land_abund_09_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_09_500_dat_xy_sf = st_as_sf(
  plot_09_500_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_09_crop <-
  st_intersection(
    plot_09_500_dat_hull_sf,
    plot_09_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_09_500$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_09 <- glmmTMB(
  abund_09 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_09_500,
  na.action = "na.fail"
)

plot_09_crop$abund_09 <- c(predict(
  mod_09,
  newdata = plot_09_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_09.2 <- ggplot() +
  geom_raster(data = plot_09_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_09)) +
  geom_contour(
    data = plot_09_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_09),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.title.x = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "VAT (°C)",
    )
heatmap_09.2
```

### heatmap_09.3

```{r}
plot_09_500_dat_hull <- climate_land_abund_09_500 %>%
  slice(chull(change3_VAT, change3_MAT))

plot_09_500_dat_hull <- plot_09_500_dat_hull %>% dplyr::select(change3_VAT, change3_MAT)

plot_09_500_dat_hull_sf <- plot_09_500_dat_hull %>%
  st_as_sf(coords = 
             c("change3_VAT",
               "change3_MAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_09_500_dat_xy <-
  with(climate_land_abund_09_500,
       expand.grid(
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300),
         change3_MAT = seq(min(change3_MAT),
                        max(change3_MAT),
                        len = 300)
       ))

plot_09_500_dat_xy_sf = st_as_sf(
  plot_09_500_dat_xy,
  coords = c("change3_VAT", "change3_MAT"))

plot_09_crop <-
  st_intersection(
    plot_09_500_dat_hull_sf,
    plot_09_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(change3_VAT = X, 
         change3_MAT = Y) %>%
  mutate(semifor_l = mean(climate_land_abund_09_500$semifor_l),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_09 <- glmmTMB(
  abund_09 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_09_500,
  na.action = "na.fail"
)

plot_09_crop$abund_09 <- c(predict(
  mod_09,
  newdata = plot_09_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_09.3 <- ggplot() +
  geom_raster(data = plot_09_crop,
              aes(y = change3_VAT,
                  x = change3_MAT,
                  fill = abund_09)) +
  geom_contour(
    data = plot_09_crop,
    aes(y = change3_VAT,
        x = change3_MAT,
        z = abund_09),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  scale_x_continuous(breaks = c(0, 1, 2)) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.title.x = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       y = "VAT (°C)",
       x = "MAT (°C)",
    )
heatmap_09.3
```


## 10. Habitat specialist, ground natural enemies

500 m:
```{r}
climate_land_abund_10_500 <-
  climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  dplyr::select(
    ts_nat_enemy_03_abund,
    sampling_effort_scaled,
    semifor_500,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_10 = ts_nat_enemy_03_abund,
    semifor_l = semifor_500
  )
```

2000 m:
```{r}
climate_land_abund_10_2000 <-
  climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  dplyr::select(
    ts_nat_enemy_03_abund,
    sampling_effort_scaled,
    semifor_2000,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_10 = ts_nat_enemy_03_abund,
    semifor_l = semifor_2000
  )
```

### heatmap_10.1

```{r}
plot_10_500_dat_hull <- climate_land_abund_10_500 %>%
  slice(chull(semifor_l, change3_VAT))

plot_10_500_dat_hull <- plot_10_500_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_10_500_dat_hull_sf <- plot_10_500_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_10_500_dat_xy <-
  with(climate_land_abund_10_500,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_10_500_dat_xy_sf = st_as_sf(
  plot_10_500_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_10_crop <-
  st_intersection(
    plot_10_500_dat_hull_sf,
    plot_10_500_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_10_500$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_10 <- glmmTMB(
  abund_10 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_10_500,
  na.action = "na.fail"
)

plot_10_crop$abund_10 <- c(predict(
  mod_10,
  newdata = plot_10_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_10.1 <- ggplot() +
  geom_raster(data = plot_10_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_10)) +
  geom_contour(
    data = plot_10_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_10),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma"
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.title.x = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.5 km",
       y = "VAT (°C)",
    )
heatmap_10.1
```

### heatmap_10.2

```{r}
plot_10_2000_dat_hull <- climate_land_abund_10_2000 %>%
  slice(chull(semifor_l, change3_VAT))

plot_10_2000_dat_hull <- plot_10_2000_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_10_2000_dat_hull_sf <- plot_10_2000_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_10_2000_dat_xy <-
  with(climate_land_abund_10_2000,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_10_2000_dat_xy_sf = st_as_sf(
  plot_10_2000_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_10_crop <-
  st_intersection(
    plot_10_2000_dat_hull_sf,
    plot_10_2000_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_10_2000$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_10 <- glmmTMB(
  abund_10 ~
    scale(semifor_l) + scale(change3_MAT) + scale(change3_VAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_10_2000,
  na.action = "na.fail"
)

plot_10_crop$abund_10 <- c(predict(
  mod_10,
  newdata = plot_10_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_10.2 <- ggplot() +
  geom_raster(data = plot_10_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_10)) +
  geom_contour(
    data = plot_10_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_10),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    breaks = c(4, 8, 12)
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.line = element_blank(),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, colour = "grey50"),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 2 km",
       y = "VAT (°C)",
    )
heatmap_10.2
```


## 11. Habitat generalist, ground natural enemies

250 m:
```{r}
climate_land_abund_11_250 <-
  climate_land_abund %>%
  filter(Nat_enemy_study == 1) %>%
  dplyr::select(
    ts_nat_enemy_04_abund,
    sampling_effort_scaled,
    semifor_250,
    change3_MAT,
    change3_VAT,
    present3_MAT,
    present3_VAT,
    StudyID_yr,
    Sampling_method,
    Block,
    Latitude,
    Longitude
  ) %>%
  ungroup() %>%
  na.omit() %>%
  mutate(
    abund_11 = ts_nat_enemy_04_abund,
    semifor_l = semifor_250
  )
```

### heatmap_11.2

```{r}
plot_11_250_dat_hull <- climate_land_abund_11_250 %>%
  slice(chull(semifor_l, change3_VAT))

plot_11_250_dat_hull <- plot_11_250_dat_hull %>% dplyr::select(semifor_l, change3_VAT)

plot_11_250_dat_hull_sf <- plot_11_250_dat_hull %>%
  st_as_sf(coords = 
             c("semifor_l",
               "change3_VAT")) %>%
  summarise() %>%
  st_cast("POLYGON") %>%
  st_convex_hull()

plot_11_250_dat_xy <-
  with(climate_land_abund_11_250,
       expand.grid(
         semifor_l = seq(min(semifor_l),
                        max(semifor_l),
                        len = 300),
         change3_VAT = seq(min(change3_VAT),
                        max(change3_VAT),
                        len = 300)
       ))

plot_11_250_dat_xy_sf = st_as_sf(
  plot_11_250_dat_xy,
  coords = c("semifor_l", "change3_VAT"))

plot_11_crop <-
  st_intersection(
    plot_11_250_dat_hull_sf,
    plot_11_250_dat_xy_sf) %>%
  st_coordinates() %>%
  as.data.frame() %>%
  rename(semifor_l = X, 
         change3_VAT = Y) %>%
  mutate(change3_MAT = mean(climate_land_abund_11_250$change3_MAT),
         StudyID_yr = NA,
         Sampling_method = NA,
         sampling_effort_scaled = 1)

mod_11 <- glmmTMB(
  abund_11 ~
    scale(semifor_l) * scale(change3_VAT) + scale(change3_MAT) + (1 | StudyID_yr:Sampling_method) + offset(log(sampling_effort_scaled)),
  family = ziGamma(link = "log"),
  ziformula = ~ 1,
  data = climate_land_abund_11_250,
  na.action = "na.fail"
)

plot_11_crop$abund_11 <- c(predict(
  mod_11,
  newdata = plot_11_crop,
  re.form = NA,
  type = "response",
  allow.new.levels = T,
  na.action = "na.pass"
))

heatmap_11.2 <- ggplot() +
  geom_raster(data = plot_11_crop,
              aes(x = semifor_l,
                  y = change3_VAT,
                  fill = abund_11)) +
  geom_contour(
    data = plot_11_crop,
    aes(x = semifor_l,
        y = change3_VAT,
        z = abund_11),
    color = 'white',
    bins = 10,
    alpha = 0.4
  ) +
  scale_fill_viridis_c(
    option = "plasma",
    alpha = 1
    ) +
  theme_classic() +
  theme(
    aspect.ratio = 1,
    # plot.title = element_text(size = 8),
    axis.text = element_text(size = 7),
    axis.title.x = element_text(size = 8, face = "bold"),
    axis.title.y = element_text(size = 8, face = "bold"),
    axis.line = element_blank(),
    legend.text = element_text(size = 7),
    legend.position = "right",
    legend.title = element_blank(),
    legend.key.width = unit(0.25, "cm"),
    legend.key.height = unit(0.25, "cm"),
    legend.box.spacing = unit(0.25, "pt"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_rect(
      colour = "black",
      linewidth = 1,
      fill = NA),
    plot.title = element_text(size = 8),
    plot.margin = unit(c(0,0,0,0), "cm")
  ) +
  labs(
    fill = "Abundance",
       x = "nat.hab (%) 0.25 km",
       y = "VAT (°C)",
    )
heatmap_11.2
```
